TST GH26807 split out test_find_replace
